STAC-23292: stskafkaexporter | make requiredAcks configurable

+++ extension/settingsproviderextension/internal/provider/kafka/kafka_settings_provider.go
@@ -170,7 +170,7 @@ func (k *SettingsProvider) checkTopicExists(ctx context.Context) error {
 
 		k.logger.Info("Topic found",
 			zap.String("topic", topic),
-			zap.Int32("partitions", int32(len(detail.Partitions))))
+			zap.Int("partitions", len(detail.Partitions)))
 	}
 
 	return nil
